+++ compressor/management/commands/compress.py
@@ -246,16 +246,12 @@ class Command(BaseCommand):
                         continue
 
                     try:
-                        result = parser.render_node(template, context, node)
+                        result = parser.render_node(template, context, node).replace(
+                            settings.COMPRESS_URL, settings.COMPRESS_URL_PLACEHOLDER
+                        )
                     except Exception as e:
                         raise CommandError("An error occurred during rendering %s: "
                                            "%s" % (template.template_name, smart_text(e)))
-
-                    if settings.COMPRESS_OFFLINE_URLLESS:
-                        result = result.replace(
-                            settings.COMPRESS_URL, settings.COMPRESS_URL_PLACEHOLDER
-                        )
-
                     offline_manifest[key] = result
                     context.pop()
                     results.append(result)
