+++ compressor/templatetags/compress.py
@@ -64,14 +64,9 @@ class CompressorMixin(object):
         key = get_offline_hexdigest(self.get_original_content(context))
         offline_manifest = get_offline_manifest()
         if key in offline_manifest:
-            result = offline_manifest[key]
-
-            if settings.COMPRESS_OFFLINE_URLLESS:
-                result = result.replace(
-                    settings.COMPRESS_URL_PLACEHOLDER, settings.COMPRESS_URL
-                )
-
-            return result
+            return offline_manifest[key].replace(
+                settings.COMPRESS_URL_PLACEHOLDER, settings.COMPRESS_URL
+            )
         else:
             raise OfflineGenerationError('You have offline compression '
                 'enabled but key "%s" is missing from offline manifest. '
